Fixed style issues as reported by IntelliJ
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1650766 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/archivers/sevenz/SevenZFile.java
@@ -163,7 +163,7 @@ public class SevenZFile implements Closeable {
         final byte archiveVersionMinor = file.readByte();
         if (archiveVersionMajor != 0) {
             throw new IOException(String.format("Unsupported 7z version (%d,%d)",
-                    Byte.valueOf(archiveVersionMajor), Byte.valueOf(archiveVersionMinor)));
+                    archiveVersionMajor, archiveVersionMinor));
         }
 
         final long startHeaderCrc = 0xffffFFFFL & Integer.reverseBytes(file.readInt());
